% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CountingDates.R
\name{difftime_leap_year}
\alias{difftime_leap_year}
\title{difftime_leap_year}
\usage{
difftime_leap_year(tfinal, tinitial, leapDatesIn = TRUE)
}
\arguments{
\item{tfinal}{Final date.}

\item{tinitial}{Initial date.}

\item{leapDatesIn}{If TRUE count leap Dates, else exclude from counting.}
}
\value{
Number of days between the specified dates.
}
\description{
Function to count the number of days between two dates.
Optional parameters to count without the leap-days.
}
\examples{
#Function accepts Dates as Dates or as characters.
difftime_leap_year(tfinal='2023-03-05',tinitial='2019-02-28',leapDatesIn=TRUE)
difftime_leap_year(tfinal=as.Date('2023-03-05'),tinitial=as.Date('2019-02-28'),leapDatesIn=TRUE)
difftime_leap_year(tfinal='2023-03-05',tinitial='2019-02-28',leapDatesIn=FALSE)
difftime_leap_year(tfinal='2023-03-05',tinitial=as.Date('2019-02-28'),leapDatesIn=FALSE)

}
\author{
Julian Chitiva and Diego Jara
}
