% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{holiDaysLDN}
\alias{holiDaysLDN}
\title{London holidays dates.}
\format{
Vector of dates.
}
\source{
Author Calculations
}
\usage{
holiDaysLDN
}
\description{
London (England) holidays dates. The holidays were created using the package bizdays.
Dates range between 1990-01-01  and 2100-12-28.

\describe{
    \item{holiDaysLDN}{Vector of dates of London holidays}
}
}
\author{
Quantil S.A.S
}
\keyword{datasets}
