% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{holiDaysNY}
\alias{holiDaysNY}
\title{New York holidays dates.}
\format{
Vector of dates.
}
\source{
Author Calculations
}
\usage{
holiDaysNY
}
\description{
New York-United States holidays dates. The holidays were created using the package timeDate.
Dates range between 1900-01-01 and 2100-12-24.

\describe{
    \item{holiDaysNY}{Vector of dates of New York holidays}
}
}
\author{
Quantil S.A.S
}
\keyword{datasets}
