% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{holiDaysNYGB}
\alias{holiDaysNYGB}
\title{New York Government Bonds holidays dates.}
\format{
Vector of dates.
}
\source{
Author Calculations
}
\usage{
holiDaysNYGB
}
\description{
New York Government Bonds -United States holidays dates. The holidays were created using the package bizdays.
Dates range between 1990-01-01 and 2100-12-24.

\describe{
    \item{holiDaysNY}{Vector of dates of New York Government Bonds holidays}
}
}
\author{
Quantil S.A.S
}
\keyword{datasets}
