% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DateManipulation.R
\name{LastDayOfMonth}
\alias{LastDayOfMonth}
\title{LastDayOfMonth}
\usage{
LastDayOfMonth(year, month, date = NULL)
}
\arguments{
\item{year}{Year as a number.}

\item{month}{Month as a number.}

\item{date}{If provided, uses year and month from this date.
It could be date or a string format date YYYY-MM-DD.}
}
\value{
Last day of the month in the current year.
}
\description{
Returns the last day of a month.
}
\examples{
# Return last day of the month in year
LastDayOfMonth(year = 2020, month = 2)

# Return last day of the month for the date
LastDayOfMonth(date = '2020-02-03')

}
\author{
Diego Jara
}
