% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DateManipulation.R
\name{NumExcel2DateR}
\alias{NumExcel2DateR}
\title{NumExcel2DateR}
\usage{
NumExcel2DateR(date)
}
\arguments{
\item{date}{numeric vector.}
}
\value{
date in R.
}
\description{
Takes a date represented by a number in Excel format (origin="1899-12-30") and returns a date in
R format.
}
\examples{
NumExcel2DateR(as.numeric(Sys.Date()))

}
\seealso{
For dates with R origin.

Other Number to Date: 
\code{\link{NumR2DateR}()}
}
\author{
Diego Jara
}
\concept{Number to Date}
