% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{wdBOG}
\alias{wdBOG}
\title{Bogota business dates.}
\format{
Vector of dates.
}
\source{
Author Calculations
}
\usage{
wdBOG
}
\description{
Bogota (Colombia) business dates. Dates range between 1998-01-02 and 2030-12-31.

\describe{
    \item{wdBOG}{Vector of dates of Bogota business days}
}
}
\author{
Quantil S.A.S
}
\keyword{datasets}
