% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{wdNY}
\alias{wdNY}
\title{New York business dates.}
\format{
Vector of dates.
}
\source{
Author Calculations
}
\usage{
wdNY
}
\description{
New York (United States) business dates. Dates range between 2000-01-03 and 2030-12-31.

\describe{
    \item{wdNY}{Vector of dates of New York business days}
}
}
\author{
Quantil S.A.S
}
\keyword{datasets}
