% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{wdNYGB}
\alias{wdNYGB}
\title{New York Government Bonds business dates.}
\format{
Vector of dates.
}
\source{
Author Calculations
}
\usage{
wdNYGB
}
\description{
New York (United States) Government Bonds business dates. Dates range between 1990-01-02 and 2100-12-31.

\describe{
    \item{wdNY}{Vector of dates of New York Government Bonds (NYGB) business days}
}
}
\author{
Quantil S.A.S
}
\keyword{datasets}
