% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{holiDaysBOG}
\alias{holiDaysBOG}
\title{Bogota holidays dates.}
\format{
Vector of dates.
}
\source{
Author Calculations
}
\usage{
holiDaysBOG
}
\description{
Bogota (Colombia) holidays dates. The holidays were created using the package timeDate.
Dates range between 2011-01-10 and 2050-12-08.

\describe{
    \item{holiDaysBOG}{Vector of dates of Bogota holidays}
}
}
\author{
Quantil S.A.S
}
\keyword{datasets}
