% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{wdLDN}
\alias{wdLDN}
\title{London business dates.}
\format{
Vector of dates.
}
\source{
Author Calculations
}
\usage{
wdLDN
}
\description{
London (England) business dates. Dates range between 2000-01-03 and 2030-12-31.

\describe{
    \item{wdLDN}{Vector of dates of London business days}
}
}
\author{
Quantil S.A.S
}
\keyword{datasets}
