% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{wdNY}
\alias{wdNY}
\title{New York Stock Exchange business dates.}
\format{
Vector of dates.
}
\source{
Author Calculations
}
\usage{
wdNY
}
\description{
New York (United States) Stock Exchange business dates. Dates range between 1990-01-02 and 2100-12-31.

\describe{
    \item{wdNY}{Vector of dates of New York Stock Exchange (NYSE) business days}
}
}
\author{
Quantil S.A.S
}
\keyword{datasets}
